remove redundant cast

+++ contracts/0.4.24/nos/NodeOperatorsRegistry.sol
@@ -706,7 +706,7 @@ contract NodeOperatorsRegistry is INodeOperatorsRegistry, AragonApp, IStakingMod
     ) external {
         require(_fromIndex <= UINT64_MAX, "FROM_INDEX_TOO_LARGE");
         /// @dev safemath(unit256) checks for overflow on addition, so _keysCount is guaranteed <= UINT64_MAX
-        require(uint256(_fromIndex).add(_keysCount) <= UINT64_MAX, "KEYS_COUNT_TOO_LARGE");
+        require(_fromIndex.add(_keysCount) <= UINT64_MAX, "KEYS_COUNT_TOO_LARGE");
         _removeUnusedSigningKeys(_nodeOperatorId, uint64(_fromIndex), uint64(_keysCount));
     }
 
@@ -730,7 +730,7 @@ contract NodeOperatorsRegistry is INodeOperatorsRegistry, AragonApp, IStakingMod
     ) external {
         require(_fromIndex <= UINT64_MAX, "FROM_INDEX_TOO_LARGE");
         /// @dev safemath(unit256) checks for overflow on addition, so _keysCount is guaranteed <= UINT64_MAX
-        require(uint256(_fromIndex).add(_keysCount) <= UINT64_MAX, "KEYS_COUNT_TOO_LARGE");
+        require(_fromIndex.add(_keysCount) <= UINT64_MAX, "KEYS_COUNT_TOO_LARGE");
         _removeUnusedSigningKeys(_nodeOperatorId, uint64(_fromIndex), uint64(_keysCount));
     }
